requirements.txt
accelerate==1.10.0
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.3
colorama==0.4.6
coloredlogs==15.0.1
diffusers==0.34.0
Django==5.2.5
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.7.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
humanfriendly==10.0
idna==3.10
imageio==2.37.0
importlib_metadata==8.7.0
Jinja2==3.1.6
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
langchain==0.3.27
langchain-core==0.3.74
langchain-huggingface==0.3.1
langchain-text-splitters==0.3.9
langsmith==0.4.14
lazy_loader==0.4
llvmlite==0.44.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numba==0.61.2
numpy==2.2.6
onnxruntime==1.22.1
opencv-python-headless==4.12.0.88
orjson==3.11.2
packaging==25.0
pillow==11.3.0
platformdirs==4.3.8
pooch==1.8.2
protobuf==6.31.1
psutil==7.0.0
pydantic==2.11.7
pydantic_core==2.33.2
PyMatting==1.1.14
pyreadline3==3.5.4
python-dotenv==1.1.1
PyYAML==6.0.2
referencing==0.36.2
regex==2025.7.34
rembg==2.0.67
requests==2.32.4
requests-toolbelt==1.0.0
rpds-py==0.27.0
safetensors==0.6.2
scikit-image==0.25.2
scipy==1.16.1
setuptools==80.9.0
sniffio==1.3.1
SQLAlchemy==2.0.43
sqlparse==0.5.3
sympy==1.14.0
tenacity==9.1.2
tifffile==2025.6.11
tokenizers==0.21.4
torch==2.8.0
tqdm==4.67.1
transformers==4.55.2
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
zipp==3.23.0
zstandard==0.23.0
